Renaming context.js to selectUserContext.js
